+++ xlators/cluster/stripe/src/stripe.c
@@ -2118,7 +2118,7 @@ stripe_create_cbk(call_frame_t *frame, void *cookie, xlator_t *this,
 
             stripe_copy_xl_array(local->fctx->xl_array, priv->xl_array,
                                  local->fctx->stripe_count);
-            inode_ctx_put(local->inode, this, (uint64_t)local->fctx);
+            inode_ctx_put(local->inode, this, (uint64_t)(uintptr_t)local->fctx);
         }
 
         /* Create itself has failed.. so return
